backend/requirements.txt
# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Database
sqlalchemy==2.0.36
pymysql==1.1.1
alembic==1.14.0
cryptography==44.0.0

# Data Validation
pydantic==2.10.5
pydantic-settings==2.7.0
python-dotenv==1.0.1

# HTTP Client
httpx==0.28.1

# Stock Data
yfinance==0.2.50

# AI/RAG
langchain==0.3.14
langchain-community==0.3.14
chromadb==0.5.23
sentence-transformers==3.3.1

# Scheduler
apscheduler==3.10.4

# Utilities
python-multipart==0.0.19
